+++ R/compare_groups.R
@@ -1,4 +1,4 @@
-#' Compare proportions for a single source across different groups
+#' Compare dietary proportions for a single source across different groups
 #' 
 #' This function takes in an object of class \code{simmr_output} and creates
 #' probabilistic comparisons for a given source and a set of at least two
@@ -13,9 +13,10 @@
 #' 
 #' @param simmr_out An object of class \code{simmr_output} created from
 #' \code{\link{simmr_mcmc}}.
-#' @param source_name The name of a source. This should match the names exactly given to \code{\link{simmr_load}}.
-#' @param group_num The integer values of the group numbers to be compared. At least two groups must be specified.
-#' @param group_name The names of the groups to be compared. At least two groups must be specified. 
+#' @param source_name The name of a source. This should match the names exactly
+#' given to \code{\link{simmr_load}}.
+#' @param groups The integer values of the group numbers to be compared. At
+#' least two groups must be specified.
 #' @param plot A logical value specifying whether plots should be produced or
 #' not.
 #' 
@@ -24,8 +25,8 @@
 #' @return If there are two groups, a vector containing the differences between
 #' the two groups proportions for that source. If there are multiple groups, a
 #' list containing the following fields: \item{Ordering }{The different
-#' possible orderings of the proportions across groups} \item{out_all
-#' }{The proportions for this source across the groups specified as
+#' possible orderings of the dietary proportions across groups} \item{out_all
+#' }{The dietary proportions for this source across the groups specified as
 #' columns in a matrix}
 #' @author Andrew Parnell <andrew.parnell@@mu.ie>
 #' @seealso See \code{\link{simmr_mcmc}} for complete examples.
@@ -124,7 +125,7 @@
 #' simmr_in
 #' 
 #' # Plot
-#' plot(simmr_in,group_num=1:8,xlab=expression(paste(delta^13, "C (\u2030)",sep="")), 
+#' plot(simmr_in,group=1:8,xlab=expression(paste(delta^13, "C (\u2030)",sep="")), 
 #'      ylab=expression(paste(delta^15, "N (\u2030)",sep="")), 
 #'      title='Isospace plot of Inger et al Geese data')
 #' 
@@ -135,33 +136,31 @@
 #' simmr_out
 #' 
 #' # Summarise output
-#' summary(simmr_out,type='quantiles',group_num=1)
-#' summary(simmr_out,type='quantiles',group_num=c(1,3))
-#' summary(simmr_out,type=c('quantiles','statistics'),group_num=c(1,3))
+#' summary(simmr_out,type='quantiles',group=1)
+#' summary(simmr_out,type='quantiles',group=c(1,3))
+#' summary(simmr_out,type=c('quantiles','statistics'),group=c(1,3))
 #' 
 #' # Plot - only a single group allowed
-#' plot(simmr_out,type='boxplot',group_num=2,title='simmr output group 2')
-#' plot(simmr_out,type=c('density','matrix'),group_num=6,title='simmr output group 6')
+#' plot(simmr_out,type='boxplot',group=2,title='simmr output group 2')
+#' plot(simmr_out,type=c('density','matrix'),grp=6,title='simmr output group 6')
 #' 
 #' # Compare groups
-#' compare_groups(simmr_out,source='Zostera',group_num=1:2)
-#' compare_groups(simmr_out,source='Zostera',group_num=1:3)
-#' compare_groups(simmr_out,source='U.lactuca',group_num=c(4:5,7,2))
+#' compare_groups(simmr_out,source='Zostera',groups=1:2)
+#' compare_groups(simmr_out,source='Zostera',groups=1:3)
+#' compare_groups(simmr_out,source='U.lactuca',groups=c(4:5,7,2))
 #' }
 #' 
 #' @export
 compare_groups = function(simmr_out,
                           source_name = simmr_out$input$source_names[1],
-                          group_num = 1:2,
-                          group_name = NULL,
+                          groups = 1:2,
                           plot = TRUE) {
   UseMethod('compare_groups')
 }
 #' @export
 compare_groups.simmr_output = function(simmr_out,
                           source_name = simmr_out$input$source_names[1],
-                          group_num = 1:2,
-                          group_name = NULL,
+                          groups = 1:2,
                           plot = TRUE) {
 
 # Function to compare between groups both via textual output and with boxplots
@@ -174,18 +173,6 @@ compare_groups.simmr_output = function(simmr_out,
 #   - provide the top most likely orderings of the groups
 # An optional boxplot of the groups
   
-  # Get the groups to use - always specify in terms of numbers
-  # Convert to group numbers if necessary
-  if(!is.null(group_name)) {
-    group_num = match(group_name, levels(simmr_out$input$group))
-    if(any(is.na(group_num))) {
-      print(data.frame(Name = unique(simmr_out$input$group), 
-                       Number = unique(simmr_out$input$group_int)))
-      stop("group_name not found in simmr input object. See table above for group names/numbers")
-    }
-  }
-  groups = group_num  
-
 # Throw an error if only one group is specified
 if(length(groups)==1) stop("Please use plot(...) or summary(...) if you just want to look at one group.")
   
@@ -212,7 +199,7 @@ if(length(groups)==2) {
     # Stupid fix for packaging ggplot things
     Group = Proportion = NULL
     df = data.frame(Proportion=c(out_all_grp_1,out_all_grp_2),Group=c(rep(paste(group_names[1]),length(out_all_grp_1)),rep(paste(group_names[2]),length(out_all_grp_2))))
-    p = ggplot(df,aes(x=Group,y=Proportion,fill=Group)) + geom_boxplot(alpha=0.5,outlier.size=0) + theme_bw() + theme(legend.position='none') + ggtitle(paste("Comparison of proportions for groups",group_names[1],'and',group_names[2],'for source',source_name))
+    p = ggplot(df,aes(x=Group,y=Proportion,fill=Group)) + geom_boxplot(alpha=0.5,outlier.size=0) + theme_bw() + theme(legend.position='none') + ggtitle(paste("Comparison of dietary proportions for groups",group_names[1],'and',group_names[2],'for source',source_name))
     print(p)
   }
   
@@ -252,7 +239,7 @@ if(length(groups)>2) {
       geom_boxplot(alpha=0.5,outlier.size=0) + 
       xlab('Group') +
       theme_bw() + theme(legend.position='none') + 
-      ggtitle(paste("Comparison of proportions for source",source_name))
+      ggtitle(paste("Comparison of dietary proportions for source",source_name))
     print(p)
   }
   
@@ -260,9 +247,9 @@ if(length(groups)>2) {
 
 # Return output
 if(length(groups)==2) {
-  invisible(list(out_diff, p))
+  invisible(list(out_diff))
 } else {
-  invisible(list(Ordering=Ordering,out_all=out_all,p=p))
+  invisible(list(Ordering=Ordering,out_all=out_all))
 }  
 
 }