+++ R/posterior_predictive.simmr_output.R
@@ -3,8 +3,7 @@
 #' This function takes the output from \code{\link{simmr_mcmc}} and plots the posterior predictive distribution to enable visualisation of model fit. The simulated posterior predicted values are returned as part of the object and can be saved for external use
 #'
 #' @param simmr_out A run of the simmr model from \code{\link{simmr_mcmc}}
-#' @param group_num Which group number(s) to plot. Defaults to first group
-#' @param group_name Which group name(s) to plot.
+#' @param group Which group to run it for (currently only numeric rather than group names)
 #' @param prob The probability interval for the posterior predictives. The default is 0.5 (i.e. 50pc intervals)
 #' @param plot_ppc Whether to create a bayesplot of the posterior predictive or not. 
 #'
@@ -47,38 +46,22 @@
 #' post_pred = posterior_predictive(simmr_1_out)
 #' }
 posterior_predictive = function(simmr_out,
-                                group_num = 1,
-                                group_name = NULL,
+                                group = 1,
                                 prob = 0.5,
                                 plot_ppc = TRUE) {
   UseMethod('posterior_predictive')
 }  
 #' @export
 posterior_predictive.simmr_output = function(simmr_out,
-                                             group_num = 1,
-                                             group_name = NULL,
+                                             group = 1,
                                              prob = 0.5,
                                              plot_ppc = TRUE) {
-  
-  # Get the groups to use - always specify in terms of numbers
-  # Convert to group numbers if necessary
-  if(!is.null(group_name)) {
-    group_num = match(group_name, levels(simmr_out$input$group))
-    if(any(is.na(group_num))) {
-      print(data.frame(Name = unique(simmr_out$input$group), 
-                       Number = unique(simmr_out$input$group_int)))
-      stop("group_name not found in simmr input object. See table above for group names/numbers")
-    }
-  }
-  group = group_num
   
 # Can't do more than 1 group for now
 if(length(group) > 1) stop("Multiple groups not supported")
 # Get the original jags script
 model_string_old = simmr_out$output[[group]]$model$model()
-
-
-
+  
 # Plug in y_pred
 copy_lines = model_string_old[6]
 copy_lines = sub("y\\[i","y_pred\\[i",copy_lines)
@@ -95,6 +78,24 @@ output = R2jags::jags(data=simmr_out$output[[group]]$model$data(),
 
 y_post_pred = output$BUGSoutput$sims.list$y_pred
 
+# Make is look nicer
+low_prob = 0.5 - prob/2
+high_prob = 0.5 + prob/2
+y_post_pred_ci = apply(y_post_pred, 
+                       2:3, 
+                       'quantile', 
+                       prob = c(low_prob, high_prob))
+y_post_pred_out = data.frame(
+  interval = matrix(y_post_pred_ci, 
+                    ncol = simmr_out$input$n_tracers,
+                    byrow = TRUE),
+  data = as.vector(simmr_out$input$mixtures[simmr_out$input$group_int == group,])
+)
+
+y_post_pred_out$outside = y_post_pred_out[,3] > y_post_pred_out[,2] | 
+  y_post_pred_out[,3] < y_post_pred_out[,1]
+prop_outside = mean(y_post_pred_out$outside)
+
 if(plot_ppc) {
   y_rep = y_post_pred
   dim(y_rep) = c(dim(y_post_pred)[1], dim(y_post_pred)[2]*dim(y_post_pred)[3])
@@ -106,10 +107,15 @@ if(plot_ppc) {
     x = rep(1:nrow(curr_mix), simmr_out$input$n_tracers),
     prob = prob,
     fatten = 1
-  )+ ggplot2::ylab("Tracer value") + ggplot2::xlab('Observation') + ggplot2::theme_bw() + ggplot2::scale_x_continuous(breaks = 1:simmr_out$input$n_obs)
+  ) + ggplot2::ylab("Tracer value") + 
+    ggplot2::xlab('Observation') + 
+    ggplot2::ggtitle(paste0(prob*100, '% posterior predictive')) + 
+    ggplot2::theme_bw() + 
+    ggplot2::scale_x_continuous(breaks = 1:simmr_out$input$n_obs)
   print(g)
 }
 # Return the simulations
-invisible(y_post_pred)
+invisible(list(table = y_post_pred_out, 
+               prop_outside = prop_outside))
 
 }