+++ R/simmr_mcmc.R
@@ -2,7 +2,8 @@
 #' Carlo (MCMC) function
 #' 
 #' This is the main function of simmr. It takes a \code{simmr_input} object
-#' created via \code{\link{simmr_load}}, runs an MCMC to determine the #' proportions, and then outputs a \code{simmr_output} object for further
+#' created via \code{\link{simmr_load}}, runs an MCMC to determine the dietary
+#' proportions, and then outputs a \code{simmr_output} object for further
 #' analysis and plotting via \code{\link{summary.simmr_output}} and
 #' \code{\link{plot.simmr_output}}.
 #' 
@@ -14,7 +15,7 @@
 #' @param simmr_in An object created via the function \code{\link{simmr_load}}
 #' @param prior_control A list of values including arguments named \code{means}
 #' and \code{sd} which represent the prior means and standard deviations of the
-#' proportions in centralised log-ratio space. These can usually be
+#' dietary proportions in centralised log-ratio space. These can usually be
 #' left at their default values unless you wish to include to include prior
 #' information, in which case you should use the function
 #' \code{\link{simmr_elicit}}.
@@ -381,12 +382,15 @@ simmr_mcmc = function(simmr_in,
 }  
 #' @export
 simmr_mcmc.simmr_input = function(simmr_in, 
-                                  prior_control=list(means=rep(0,simmr_in$n_sources),
-                                                     sd=rep(1,simmr_in$n_sources)), 
-                                  mcmc_control=list(iter=10000,
-                                                    burn=1000,
-                                                    thin=10,
-                                                    n.chain=4)) {
+                      prior_control=list(means=rep(0,simmr_in$n_sources),
+                                         sd=rep(1,simmr_in$n_sources)), 
+                      mcmc_control=list(iter=10000,
+                                        burn=1000,
+                                        thin=10,
+                                        n.chain=4)) {
+
+# Main function to run simmr through JAGS
+# if(class(simmr_in)!='simmr_input') stop("Input argument simmr_in must have come from simmr_load")
 
 # Throw warning if n.chain =1
 if(mcmc_control$n.chain==1) warning("Running only 1 MCMC chain will cause an error in the convergence diagnostics")
@@ -418,7 +422,6 @@ model{
 }
 "
 
-# Holder to store output
 output = vector('list',length=simmr_in$n_groups)
 names(output) = levels(simmr_in$group)
 
@@ -438,7 +441,7 @@ for(i in 1:simmr_in$n_groups) {
   }
   
   # Create data object
-  data = with(simmr_in, list(
+  data = with(simmr_in,list(
     y=curr_mix,
     s_mean=source_means,
     s_sd=source_sds,