+++ R/summary.simmr_output_tdf.R
@@ -8,7 +8,7 @@
 #' 'correlations' which produces correlations between the parameters.
 #' 
 #' The quantile output allows easy calculation of 95 per cent credible
-#' intervals of the posterior proportions. The Gelman diagnostic values should be close to 1 to ensure satisfactory convergence.
+#' intervals of the posterior dietary proportions. The Gelman diagnostic values should be close to 1 to ensure satisfactory convergence.
 #' 
 #' Multiple groups are not currently supported for estimating TDFs.
 #' 
@@ -109,7 +109,7 @@ summary.simmr_output_tdf =
     if ('diagnostics'%in%type) {
       # Print out gelman diagnostics of the output
       cat('Gelman diagnostics - these values should all be close to 1.\n')
-      cat('If any are larger than 1.1, try a longer run of simmr_mcmc_tdf.\n')
+      cat('If not, try a longer run of simmr_mcmc_tdf.\n')
       print(round(out_bgr,2))
     }