+++ tests/testthat/test_simmr_print_summary.R
@@ -2,25 +2,17 @@ context("simmr print and summary")
 
 library(simmr)
 
-mix = matrix(c(-10.13, -10.72, -11.39, -11.18, -10.81, -10.7, -10.54, 
-               -10.48, -9.93, -9.37, 11.59, 11.01, 10.59, 10.97, 11.52, 11.89, 
-               11.73, 10.89, 11.05, 12.3), ncol=2, nrow=10)
-colnames(mix) = c('d13C','d15N')
-s_names=c('Source A','Source B','Source C','Source D')
-s_means = matrix(c(-14, -15.1, -11.03, -14.44, 3.06, 7.05, 13.72, 5.96), ncol=2, nrow=4)
-s_sds = matrix(c(0.48, 0.38, 0.48, 0.43, 0.46, 0.39, 0.42, 0.48), ncol=2, nrow=4)
-c_means = matrix(c(2.63, 1.59, 3.41, 3.04, 3.28, 2.34, 2.14, 2.36), ncol=2, nrow=4)
-c_sds = matrix(c(0.41, 0.44, 0.34, 0.46, 0.46, 0.48, 0.46, 0.66), ncol=2, nrow=4)
-conc = matrix(c(0.02, 0.1, 0.12, 0.04, 0.02, 0.1, 0.09, 0.05), ncol=2, nrow=4)
+data(geese_data_day1)
 
 # Load into simmr
-simmr_1 = simmr_load(mixtures=mix,
-                     source_names=s_names,
-                     source_means=s_means,
-                     source_sds=s_sds,
-                     correction_means=c_means,
-                     correction_sds=c_sds,
-                     concentration_means = conc)
+simmr_1 = with(geese_data_day1, 
+               simmr_load(mixtures=mixtures,
+                          source_names=source_names,
+                          source_means=source_means,
+                          source_sds=source_sds,
+                          correction_means=correction_means,
+                          correction_sds=correction_sds,
+                          concentration_means = concentration_means))
 # MCMC run
 simmr_1_out = simmr_mcmc(simmr_1,
                          mcmc_control = list(iter = 100, 
@@ -41,3 +33,46 @@ test_that('summary.simmr_output', {
   expect_output(summary(simmr_1_out,type='diagnostics'))
   expect_output(summary(simmr_1_out))
 })
+
+
+# Group version -----------------------------------------------------------
+
+data(geese_data)
+simmr_groups = with(geese_data, 
+                    simmr_load(mixtures=mixtures[,1:2],
+                               source_names=source_names,
+                               source_means=source_means,
+                               source_sds=source_sds,
+                               correction_means=correction_means,
+                               correction_sds=correction_sds,
+                               concentration_means = concentration_means,
+                               group=as.factor(paste('period', 
+                                                     mixtures[,3]))))
+simmr_groups_out = simmr_mcmc(simmr_groups, 
+                              mcmc_control = list(iter = 100, 
+                                                  burn = 10, 
+                                                  thin = 1, 
+                                                  n.chain = 2))
+# Taken from the simmr_mcmc example
+test_that('print.simmr_input', {
+  expect_output(print(simmr_groups))
+})
+
+test_that('print.simmr_output', {
+  expect_output(print(simmr_groups_out))
+})
+
+test_that('summary.simmr_output', {
+  expect_output(summary(simmr_groups_out,type='diagnostics'))
+  expect_output(summary(simmr_groups_out))
+})
+
+test_that('summary.simmr_output groups ', {
+  expect_output(summary(simmr_groups_out,type='statistics', group = 1))
+  expect_output(summary(simmr_groups_out,type='statistics', group = c(1, 3)))
+  summ_1 = summary(simmr_groups_out,type='statistics', group = 1)
+  summ_2 = summary(simmr_groups_out,type='statistics', group = 2)
+  expect_false(summ_1$statistics[[1]][1,1] == summ_2$statistics[[1]][1,1])
+})
+
+